requirements.txt
numba==0.55.1
numpy==1.21.4
opencv_python==4.5.5.62
Pillow==9.0.1
pygame==2.1.0
